# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MesaFácil API - Plataforma de gestão de eventos e mapas de assentos
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export type EventType = typeof EventType[keyof typeof EventType];


export const EventType = {
  wedding: 'wedding',
  birthday: 'birthday',
  graduation: 'graduation',
  corporate: 'corporate',
  other: 'other',
} as const;

export interface Event {
  id: number;
  name: string;
  date: string;
  type: EventType;
  /** @nullable */
  venue?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type EventInputType = typeof EventInputType[keyof typeof EventInputType];


export const EventInputType = {
  wedding: 'wedding',
  birthday: 'birthday',
  graduation: 'graduation',
  corporate: 'corporate',
  other: 'other',
} as const;

export interface EventInput {
  /** @minLength 1 */
  name: string;
  date: string;
  type: EventInputType;
  venue?: string;
  notes?: string;
}

export type EventUpdateType = typeof EventUpdateType[keyof typeof EventUpdateType];


export const EventUpdateType = {
  wedding: 'wedding',
  birthday: 'birthday',
  graduation: 'graduation',
  corporate: 'corporate',
  other: 'other',
} as const;

export interface EventUpdate {
  /** @minLength 1 */
  name?: string;
  date?: string;
  type?: EventUpdateType;
  /** @nullable */
  venue?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface EventStats {
  totalGuests: number;
  assignedGuests: number;
  unassignedGuests: number;
  checkedInGuests: number;
  totalSeats: number;
  occupiedSeats: number;
  availableSeats: number;
}

export interface Guest {
  id: number;
  eventId: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  group?: string | null;
  /** @nullable */
  notes?: string | null;
  checkedIn: boolean;
  /** @nullable */
  checkedInAt?: string | null;
  /** @nullable */
  floorItemId?: number | null;
  /** @nullable */
  seatNumber?: number | null;
  createdAt: string;
}

export interface GuestInput {
  /** @minLength 1 */
  name: string;
  phone?: string;
  group?: string;
  notes?: string;
}

export interface GuestUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  group?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  floorItemId?: number | null;
  /** @nullable */
  seatNumber?: number | null;
}

export type FloorItemType = typeof FloorItemType[keyof typeof FloorItemType];


export const FloorItemType = {
  'round-table': 'round-table',
  'rectangle-table': 'rectangle-table',
  'square-table': 'square-table',
  'couple-table': 'couple-table',
  'buffet-table': 'buffet-table',
  stage: 'stage',
  bathroom: 'bathroom',
  'dj-booth': 'dj-booth',
  entrance: 'entrance',
  'emergency-exit': 'emergency-exit',
} as const;

export interface FloorItem {
  id: number;
  eventId: number;
  type: FloorItemType;
  /** @nullable */
  label?: string | null;
  x: number;
  y: number;
  width: number;
  height: number;
  rotation: number;
  capacity: number;
  createdAt: string;
}

export type FloorItemInputType = typeof FloorItemInputType[keyof typeof FloorItemInputType];


export const FloorItemInputType = {
  'round-table': 'round-table',
  'rectangle-table': 'rectangle-table',
  'square-table': 'square-table',
  'couple-table': 'couple-table',
  'buffet-table': 'buffet-table',
  stage: 'stage',
  bathroom: 'bathroom',
  'dj-booth': 'dj-booth',
  entrance: 'entrance',
  'emergency-exit': 'emergency-exit',
} as const;

export interface FloorItemInput {
  type: FloorItemInputType;
  label?: string;
  x: number;
  y: number;
  width: number;
  height: number;
  rotation?: number;
  capacity: number;
}

export type FloorItemUpdateType = typeof FloorItemUpdateType[keyof typeof FloorItemUpdateType];


export const FloorItemUpdateType = {
  'round-table': 'round-table',
  'rectangle-table': 'rectangle-table',
  'square-table': 'square-table',
  'couple-table': 'couple-table',
  'buffet-table': 'buffet-table',
  stage: 'stage',
  bathroom: 'bathroom',
  'dj-booth': 'dj-booth',
  entrance: 'entrance',
  'emergency-exit': 'emergency-exit',
} as const;

export interface FloorItemUpdate {
  type?: FloorItemUpdateType;
  /** @nullable */
  label?: string | null;
  x?: number;
  y?: number;
  width?: number;
  height?: number;
  rotation?: number;
  capacity?: number;
}