# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
  description: MesaFácil API - Plataforma de gestão de eventos e mapas de assentos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{eventId}/stats:
    get:
      operationId: getEventStats
      tags: [events]
      summary: Get event seating statistics
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventStats"
        "404":
          description: Not found
...
components:
  schemas:
...
    Event:
      type: object
      required: [id, name, date, type, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        date:
          type: string
          format: date
        type:
          type: string
          enum: [wedding, birthday, graduation, corporate, other]
        venue:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    EventInput:
      type: object
      required: [name, date, type]
      properties:
        name:
          type: string
          minLength: 1
        date:
          type: string
          format: date
        type:
          type: string
          enum: [wedding, birthday, graduation, corporate, other]
        venue:
          type: string
        notes:
          type: string
    EventUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        date:
          type: string
          format: date
        type:
          type: string
          enum: [wedding, birthday, graduation, corporate, other]
        venue:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    EventStats:
      type: object
      required: [totalGuests, assignedGuests, unassignedGuests, checkedInGuests, totalSeats, occupiedSeats, availableSeats]
      properties:
        totalGuests:
          type: integer
        assignedGuests:
          type: integer
        unassignedGuests:
          type: integer
        checkedInGuests:
          type: integer
        totalSeats:
          type: integer
        occupiedSeats:
          type: integer
        availableSeats:
          type: integer
...